move placement service to container.api bundle
