--- org.opentosca.container.api/src/org/opentosca/container/api/dto/request/StartPlacementRequest.java
@@ -1,25 +1,0 @@
-package org.opentosca.container.api.dto.request;
-
-import javax.xml.bind.annotation.XmlAccessType;
-import javax.xml.bind.annotation.XmlAccessorType;
-import javax.xml.bind.annotation.XmlRootElement;
-import javax.xml.bind.annotation.XmlValue;
-
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-
-@XmlRootElement(name = "correlationID")
-@XmlAccessorType(XmlAccessType.FIELD)
-@JsonIgnoreProperties(ignoreUnknown = true)
-public class StartPlacementRequest {
-
-    @XmlValue
-    private String correlationId;
-
-    public String getCorrelationId() {
-        return this.correlationId;
-    }
-
-    public void setCorrelationId(final String correlationId) {
-        this.correlationId = correlationId;
-    }
-}
