+++ org.opentosca.placement/OSGI-INF/PlacementService.xml
